Update network.sh
If the VM_NET_TAP is there from a prior container remove it so a new one may be built in its place. Its annoying when doing a docker-compose up and it fails because the old container leaves an interface and this prevents it from halting or looping for an error.