feat: allow ip not exist
In CNI, MACVLAN can be used only as a Layer 2 network without assigning an IP address. In this case, the container will not be assigned an IP address, but the VM will obtain an IP address using DHCP. Therefore, it is necessary to allow the IP address to be empty.